MAINT: rename the package/module
