+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "scpdt"
+name = "scipy_doctest"
 maintainers = [
     {name = "SciPy developers", email = "scipy-dev@python.org"}
     ]
@@ -38,4 +38,4 @@ addopts = "--verbose --color=yes"
 # specify project entrypoints to make each plugin discoverable by pytest
 # ref: https://docs.pytest.org/en/latest/how-to/writing_plugins.html#making-your-plugin-installable-by-others
 [project.entry-points.pytest11]
-scpdt = "scpdt.plugin"
+scipy_doctest = "scipy_doctest.plugin"
