+++ scipy_doctest/__init__.py
