+++ scipy_doctest/__main__.py
