+++ scipy_doctest/conftest.py
