+++ scipy_doctest/impl.py
