+++ scipy_doctest/plugin.py
