+++ scipy_doctest/tests/failure_cases.py
