+++ scipy_doctest/tests/failure_cases_2.py
