+++ scipy_doctest/tests/finder_cases.py
