+++ scipy_doctest/tests/module_cases.py
