+++ scipy_doctest/tests/scipy_ndimage_tutorial_clone.rst
