+++ scipy_doctest/tests/stopwords_cases.py
