+++ scipy_doctest/tests/test_parser.py
