+++ scipy_doctest/tests/test_runner.py
