+++ scipy_doctest/tests/test_skipmarkers.py
