+++ scipy_doctest/tests/test_testmod.py
@@ -104,7 +104,7 @@ class TestLocalFiles:
         # A doctest tries to open a local file. Test that it works
         # (internally, the file will need to be copied).
         config = DTConfig()
-        config.local_resources = {'scpdt.tests.local_file_cases.local_files':
+        config.local_resources = {'scipy_doctest.tests.local_file_cases.local_files':
                                   ['local_file.txt']}
         res, _ = _testmod(local_file_cases, config=config,
                          strategy=[local_file_cases.local_files],
@@ -117,7 +117,7 @@ class TestLocalFiles:
     def test_sio_octave(self):
         # scipy/tutorial/io.rst : octave_a.mat file
         config = DTConfig()
-        config.local_resources = {'scpdt.tests.local_file_cases.sio':
+        config.local_resources = {'scipy_doctest.tests.local_file_cases.sio':
                                   ['octave_a.mat']}
         res, _ = _testmod(local_file_cases, config=config,
                           strategy=[local_file_cases.sio],
