+++ scipy_doctest/util.py
